Refactor
- Merge Level and LevelContainer (actions, reducers, components) into Game/GameContainer.
- Factor out Game from GameContainer.
- Refactor structure of calculator: Now calculator is separated from game.
- Extract navigator from index.js
- Remove appReducers (add into index.js)
