+++ src/components/game/LevelSelection.js
@@ -3,9 +3,17 @@ import {View, Text, TouchableOpacity} from "react-native";
 import {Container, Content, List, ListItem} from "native-base";
 import {LevelEfficacyStars} from "../common/LevelEfficacyStars";
 import {LEVEL_SELECTION_STYLES} from "../../styles/game/styles"
-import {LEVEL_EFFICACY_STARS_STYLES} from "../../styles/common/styles"
+import {HEADER_STYLES, LEVEL_EFFICACY_STARS_STYLES} from "../../styles/common/styles"
+import {BackButton} from "../common/BackButton";
 
 export default class LevelSelection extends React.Component {
+    static navigationOptions = ({navigation, screenProps}) => ({
+        title: 'ARCADE',
+        headerLeft: <BackButton goBack={navigation.goBack}/>,
+        headerStyle: HEADER_STYLES.header,
+        headerTitleStyle: HEADER_STYLES.title,
+    });
+
     constructor(props) {
         super(props);
 
@@ -13,12 +21,12 @@ export default class LevelSelection extends React.Component {
     }
 
     goToLevel(levelNumber) {
-        this.props.navigation.navigate('Level', {
+        this.props.navigation.navigate('PlayLevel', {
             levelNumber: levelNumber
         });
     }
 
-     render(){
+    render() {
         return (
             <Content style={LEVEL_SELECTION_STYLES.list}>
                 <TouchableOpacity style={LEVEL_SELECTION_STYLES.listItem}>
@@ -45,5 +53,5 @@ export default class LevelSelection extends React.Component {
                 </TouchableOpacity>
             </Content>
         )
-     }
+    }
 }
