+++ src/components/game/calculator/Calculator.js
@@ -1,13 +1,7 @@
 import React from 'react';
 import {View} from 'react-native';
-
-import UserAnswerFeedback from './UserAnswerFeedback';
-import Header from './header/Header';
-import CountdownBar from './CountdownBar';
 import OperationDisplay from './OperationDisplay';
 import CalculatorKeyboard from './keyboard/CalculatorKeyboard';
-
-import {CALCULATOR_STYLES} from "../../../styles/game/calculator/styles";
 
 export default class Calculator extends React.Component {
     constructor(props) {
@@ -16,10 +10,7 @@ export default class Calculator extends React.Component {
 
     render() {
         return (
-            <View style={CALCULATOR_STYLES.calculator}>
-                <UserAnswerFeedback {...this.props}/>
-                <Header {...this.props}/>
-                <CountdownBar {...this.props}/>
+            <View>
                 <OperationDisplay {...this.props}/>
                 <CalculatorKeyboard {...this.props}/>
             </View>
