+++ src/components/game/header/CalculationTimer.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {View, Text} from "react-native";
 
-import {HEADER_STYLES} from "../../../../styles/game/calculator/header/styles";
+import {HEADER_STYLES} from "../../../styles/game/calculator/header/styles";
 
 
 export default class CalculationTimer extends React.Component {
