+++ src/reducers/game_reducer.js
@@ -1,6 +1,7 @@
 import {
-    ARCADE_ERASE_INPUT, ARCADE_TYPE_INPUT, CREATE_TRIAL, HIDE_FEEDBACK, SHOW_FEEDBACK
-} from '../actions/calculator_actions'
+    ARCADE_ERASE_INPUT, ARCADE_TYPE_INPUT, CREATE_TRIAL, FINISH_LEVEL, HIDE_FEEDBACK, SHOW_FEEDBACK, START_LEVEL,
+    SUBMIT_TRIAL
+} from '../actions/game_actions'
 
 const MAX_NUMBER_OF_DIGITS = 8;
 
@@ -23,7 +24,14 @@ const initialState = {
     },
     header: {
         visible: true,
-    }
+    },
+    level: 1,
+    trials: [],
+    levelFinished: true,
+    totalTrials: 20,
+    totalCorrect: 0,
+    efficacy: 0,
+    maxTimeForCountdownInMs: 30000,
 };
 
 const updateUserInput = function (actualInput, newInput) {
@@ -42,7 +50,7 @@ const updateUserInput = function (actualInput, newInput) {
     return updatedInput;
 };
 
-export function calculatorReducer(state = initialState, action) {
+export function gameReducer(state = initialState, action) {
     switch (action.type) {
         case CREATE_TRIAL:
             return {
@@ -94,6 +102,34 @@ export function calculatorReducer(state = initialState, action) {
                 header: {
                     visible: true,
                 },
+            };
+
+        case START_LEVEL:
+            return {
+                ...state,
+                trials: initialState.trials,
+                levelFinished: false,
+                level: action.level,
+                totalTrials: initialState.totalTrials,
+                totalCorrect: initialState.totalCorrect,
+                efficacy: initialState.efficacy,
+                maxTimeForCountdownInMs: initialState.maxTimeForCountdownInMs,
+            };
+
+        case SUBMIT_TRIAL:
+            return {
+                ...state,
+                trials: state.trials.concat(action.trial),
+                totalCorrect: action.trial.input === action.trial.operation.result
+                    ? state.totalCorrect + 1
+                    : state.totalCorrect
+            };
+
+        case FINISH_LEVEL:
+            return {
+                ...state,
+                levelFinished: true,
+                efficacy: (state.totalCorrect / state.totalTrials) * 100
             };
         default:
             return state
